rpi/audio_streamer/requirements.txt
pyaudio==0.2.14
websockets==14.1
aiohttp==3.11.11
numpy==1.26.4
evdev==1.7.1
selenium==4.27.1
Pillow==11.0.0
